Move the runtime tests into their own file.
